test: migrate shell specs to main process (#35058)
